Implementing fix for POST request retry issue
If a POST request is sent when the Datera box is under extremely heavy
load, there is a scenario where a reply may not have been recieved
before the underlying connection is dropped.  The retry code tries again
and recieves a ConflictError in this case which was previously being
propagated up to the user.

This patch fixes the issue by catching that ConflictError (only for POST
requests) and translating the POST request to a GET request for
subsequent polling until a response is recieved.  This response is then
processed and returned normally.