Added filter param accumulation skip
This only kicks in if the product version is < 3.3.0.0 since there
is a fix present in this version of the product.

Essentially an uneccessary accumulation operation is performed when a
"filter" parameter is provided and the total number of AppInstances on
the product exceeds 100.

Moved datera version comparison into connection.py